lus won't give us that anymore

+++ OTRExporter/BlobExporter.cpp
@@ -5,7 +5,7 @@ void OTRExporter_Blob::Save(ZResource* res, const fs::path& outPath, BinaryWrite
 {
 	ZBlob* blob = (ZBlob*)res;
 	
-	WriteHeader(blob, outPath, writer, LUS::ResourceType::Blob);
+	WriteHeader(blob, outPath, writer, static_cast<uint32_t>(LUS::ResourceType::Blob));
 
 	auto start = std::chrono::steady_clock::now();
